// src/app/modulos/vistas-generales/interfaces/JugadorTirosLibres.ts
export interface JugadorTirosLibres {
  posicion: number;
  tirosLibres: number;
  equipo: string;
  nombre: string;
}

export interface JugadorTiros2Puntos{
  posicion: number;
  tiros2Puntos: number;
  equipo: string;
  nombre: string;
}

export interface JugadorTiros3Puntos{
  posicion: number;
  tiros3Puntos: number;
  equipo: string;
  nombre: string;
}

export interface JugadorAsistencias{
  posicion: number;
  asistencias: number;
  equipo: string;
  nombre: string;
}